fixed a lot of compile time errors

+++ src/Entropy.c
@@ -14,7 +14,7 @@
 double calculateEntropy( char* data, int size )
 {
     // Calculate entropy for given data and return it.
-    int *freq = calculateFrequency(data,size);
+    long *freq = calculateFrequency(data,size);
     double length = size;
     double *prob = (double*)malloc(sizeof(double)*256);
     double entropy;
@@ -30,7 +30,7 @@ double calculateEntropy( char* data, int size )
 
 long int* calculateFrequency(char* data,int size)
 {
-	long int *freq = (int*)malloc(sizeof(int)*256);
+	long *freq = (long*)malloc(sizeof(long)*256);
     int i;
     for(i=0;i<256;i++)
     {
